`partitionMap` is now part of `http.util.Collections.SeqOps`
